+++ app/src/main/res/values/strings.xml
@@ -5,19 +5,57 @@
     <string name="dialog_no_valid_email">유효한 이메일 형식이 아닙니다.</string>
     <string name="dialog_no_valid_password">비밀번호는 8~20자 이내로 영문 대소문자, 숫자, 특수문자 중 3가지 이상 혼용해야 합니다.</string>
 
+    <!-- common -->
     <string name="back">뒤로가기</string>
     <string name="wavve_logo">wavve 로고</string>
+    <string name="ok">확인</string>
+    <string name="info">도움말</string>
+    <string name="more">더보기</string>
+
+    <!-- auth -->
+    <string name="sign_up">회원가입</string>
+    <string name="sign_up_button">Wavve 회원가입</string>
     <string name="email_or_id">이메일 주소 또는 아이디</string>
     <string name="login">로그인</string>
     <string name="password">비밀번호</string>
+    <string name="find_id">아이디 찾기</string>
     <string name="hide">hide</string>
     <string name="show">show</string>
-    <string name="find_id">아이디 찾기</string>
     <string name="reset_password">비밀번호 재설정</string>
-    <string name="sign_up">회원가입</string>
     <string name="social_description">또는 다른 서비스 계정으로 로그인</string>
     <string name="social_info">SNS계정으로 간편하게 가입하여 서비스를 이용하실 수 있습니다.\n기존 POOQ 계정 또는 Wavve 계정과는 연동되지 않으니 이용에 참고하세요</string>
+    <string name="placeholder_email">wavve@example.com</string>
+    <string name="placeholder_password">Wavve 비밀번호 설정</string>
+    <string name="helper_text_password">비밀번호는 8~20자 이내로 영문 대소문자, 숫자, 특수문자 중 3가지 이상 혼용하여 입력해 주세요.</string>
+    <string name="social_description_2">또는 다른 서비스 계정으로 가입</string>
+    <string name="email_and_password">이메일과 비밀번호</string>
+    <string name="only">만으로</string>
+    <string name="wavve_enjoy">Wavve를 즐길 수</string>
+    <string name="able"> 있어요!</string>
+    <string name="fail_to_login">로그인에 실패했습니다.</string>
 
+    <!-- Home -->
+    <string name="welcome">환영합니다!</string>
+    <string name="home_banner">배너 이미지</string>
+    <string name="editor_recommended_work">에디터 추천작</string>
+    <string name="today_top">Top 20 이미지</string>
+    <string name="editor_recommend_title">믿고 보는 웨이브 에디터 추천작</string>
+    <string name="today_top_20_title">오늘의 TOP 20</string>
+    <string name="drama">드라마</string>
+    <string name="movie">영화</string>
+    <string name="animation">애니</string>
+    <string name="new_classic">뉴클래식</string>
+    <string name="overseas_series">해외시리즈</string>
+    <string name="variety_show">예능</string>
+    <string name="wifi">WIFI</string>
+    <string name="live">LIVE</string>
+    <string name="home">홈</string>
+
+    <!-- Search -->
+    <string name="search">검색</string>
+
+    <!-- My -->
+    <string name="my">MY</string>
     <string name="profile">프로필</string>
     <string name="notifications">알림</string>
     <string name="settings">설정</string>
@@ -28,38 +66,4 @@
     <string name="no_viewing_history">시청내역이 없어요.</string>
     <string name="interested_program">관심 프로그램</string>
     <string name="no_interested_program">관심 프로그램이 없어요.</string>
-
-    <string name="placeholder_email">wavve@example.com</string>
-    <string name="helper_text_email">로그인, 비밀번호 찾기, 알림에 사용되니 정확한 이메일을 입력해주세요.</string>
-    <string name="placeholder_password">Wavve 비밀번호 설정</string>
-    <string name="helper_text_password">비밀번호는 8~20자 이내로 영문 대소문자, 숫자, 특수문자 중 3가지 이상 혼용하여 입력해 주세요.</string>
-    <string name="sign_up_button">Wavve 회원가입</string>
-    <string name="ok">확인</string>
-    <string name="info">도움말</string>
-    <string name="social_description_2">또는 다른 서비스 계정으로 가입</string>
-
-    <string name="email_and_password">이메일과 비밀번호</string>
-    <string name="only">만으로</string>
-    <string name="wavve_enjoy">Wavve를 즐길 수</string>
-    <string name="able"> 있어요!</string>
-    <string name="fail_to_login">로그인에 실패했습니다.</string>
-    <string name="welcome">환영합니다!</string>
-    <string name="home_banner">배너 이미지</string>
-    <string name="editor_recommended_work">에디터 추천작</string>
-    <string name="today_top">Top 20 이미지</string>
-    <string name="editor_recommend_title">믿고 보는 웨이브 에디터 추천작</string>
-    <string name="today_top_20_title">오늘의 TOP 20</string>
-    <string name="more">더보기</string>
-    <string name="drama">드라마</string>
-    <string name="movie">영화</string>
-    <string name="animation">애니</string>
-    <string name="new_classic">뉴클래식</string>
-    <string name="overseas_series">해외시리즈</string>
-    <string name="variety_show">예능</string>
-    <string name="wifi">WIFI</string>
-    <string name="live">LIVE</string>
-    <string name="home">홈</string>
-    <string name="search">검색</string>
-    <string name="my">MY</string>
-
 </resources>
